% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_doi}
\alias{check_doi}
\title{Check for valid DOI}
\usage{
check_doi(doi)
}
\arguments{
\item{doi}{Digital Object Identifier (string)}
}
\value{
Boolean
}
\description{
This helper function checks whether a DOI is valid. The regular expression
is based on the one provided by CrossRef as providing the highest coverage
(\url{https://bit.ly/doi-regex}).
}
\examples{
  check_doi(doi = '10.1002/job.1787')
  check_doi(doi = '10.1111/j.1365-2044.2012.07128.x')
}
