% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_or_url}
\alias{construct_or_url}
\title{OpenRetractions URL}
\usage{
construct_or_url(doi)
}
\arguments{
\item{doi}{Digital Object Identifier (string)}
}
\value{
URL (string)
}
\description{
Helper function to easily maintain the API calls to
\url{http://openretractions.com}.
}
