% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_doi}
\alias{find_doi}
\title{Find DOIs in strings}
\usage{
find_doi(strings)
}
\arguments{
\item{strings}{Vector of strings to check for DOIs}
}
\value{
Vector of DOIs
}
\description{
Helper function to find DOIs in strings. Can occasionally erroneously
extract DOIs (subset of another DOI for example). Regular expression
based on CrossRef (\url{http://bit.ly/doi-regex}; see also 
\code{\link{check_doi}}).
}
\examples{
  find_doi('This contains 10.1111/j.1365-2044.2012.07128.x')
}
