% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_date}
\alias{get_date}
\title{Convert timestamp to Date}
\usage{
get_date(timestamp)
}
\arguments{
\item{timestamp}{Numeric timestamp}
}
\value{
a Date
}
\description{
Helper function to convert a timestamp from openretractions.com into a date.
}
