% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractcheck.R
\name{query_or}
\alias{query_or}
\title{Query \url{http://openretractions.com} for retractions}
\usage{
query_or(doi)
}
\arguments{
\item{doi}{Character. A digital object identifier (DOI).}
}
\value{
Data frame
}
\description{
Using the \url{http://openretractions.com} API, this function
checks whether a DOI has been updated, when that update was made,
and what type of update was made.
}
\examples{
\donttest{
  query_or('10.1002/job.1787')
}
}
