% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractcheck.R
\name{query_rw}
\alias{query_rw}
\title{Query \url{http://retractiondatabase.org} for retractions}
\usage{
query_rw(doi)
}
\arguments{
\item{doi}{Character. A digital object identifier (DOI).}
}
\description{
Using \url{http://retractiondatabase.org}, this function
checks whether a DOI has been updated, when that update was made,
and what type of update was made.
}
\examples{
\dontrun{
  query_or('10.1002/job.1787')
  query_rw('10.1002/job.1787')
}
}
