% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractcheck.R
\docType{package}
\name{retractcheck}
\alias{retractcheck}
\alias{retractcheck-package}
\title{retractcheck: Retraction scanner}
\usage{
retractcheck(dois, database = c("or", "rw"))
}
\arguments{
\item{dois}{Vector of strings containing only DOIs}

\item{database}{Character. Abbreviation of the databases to search
(\code{or} for openretractions.com and \code{rw} for
retractiondatabase.com). Note that in the absence of an API,
searching retractiondatabase.com is rather slow.}
}
\value{
\code{\link{retractcheck}} dataframe
}
\description{
Using 'Digital Object Identifiers', check for retracted (or otherwise
updated) articles using 'Open Retractions' <http://openretractions.com>.

Using the \url{http://openretractions.com} API, this function
checks whether a DOI has been updated, when that update was made,
and what type of update was made.
}
\examples{
\dontrun{
  retractcheck(c('10.1002/job.1787',
                 '10.1111/j.1365-2044.2012.07128.x'))
}
}
