% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractcheck.R
\name{retractcheck_dir}
\alias{retractcheck_dir}
\title{Check files in directory for retractions}
\usage{
retractcheck_dir(path)
}
\arguments{
\item{path}{Path to directory to check}
}
\value{
\code{\link{retractcheck}} dataframe
}
\description{
Check all HTML, DOCX, PDF, and RTF files in a directory for updates to 
referenced DOIs.
}
\examples{
\dontrun{
  retractcheck_dir(path = '.')
}
}
