% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractcheck.R
\name{retractcheck_docx}
\alias{retractcheck_docx}
\title{Check docx file for retractions}
\usage{
retractcheck_docx(path, ...)
}
\arguments{
\item{path}{Path to DOCX file to check}

\item{...}{Arguments passed on to \code{retractcheck}
\describe{
  \item{database}{Character. Abbreviation of the databases to search
(\code{or} for openretractions.com).}
  \item{return}{Character. If \code{all}, all databases are queried and all
results are returned; if \code{unique}, the databases are queried in
the order specified in \code{database} until either a correction or
retraction notice is found or all databases have been queried.}
}}
}
\value{
\code{\link{retractcheck}} dataframe without filenames
}
\description{
Check a DOCX file for retractions.
}
\examples{
\donttest{
  retractcheck_docx('manuscript.docx')
}
}
