% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractcheck.R
\name{retractcheck_html}
\alias{retractcheck_html}
\title{Check html file for retractions}
\usage{
retractcheck_html(path)
}
\arguments{
\item{path}{Path to html file to check}
}
\value{
\code{\link{retractcheck}} dataframe
}
\description{
Check a html file for retractions.
}
\examples{
\dontrun{
  retractcheck_html('manuscript.html')
}
}
