% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractcheck.R
\name{retractcheck_pdf}
\alias{retractcheck_pdf}
\title{Check pdf file for retractions}
\usage{
retractcheck_pdf(path)
}
\arguments{
\item{path}{Path to pdf file to check}
}
\value{
\code{\link{retractcheck}} dataframe without filenames
}
\description{
Check a pdf file for retractions.
}
\examples{
\dontrun{
  retractcheck_pdf('manuscript.pdf')
}
}
