% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_date}
\alias{get_date}
\title{Convert timestamp to Date}
\usage{
get_date(x, database = "or")
}
\arguments{
\item{x}{Numeric or Character. Numeric timestamp from openretractions.com or character from retractiondatabase.com}

\item{database}{Character. Abbreviation of the database (either \code{or} for openretractions.com, or \code{rw} for retractiondatabase.com)}
}
\value{
a Date
}
\description{
Helper function to convert a timestamp from openretractions.com into a date.
}
