% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractcheck.R
\name{retractcheck_docx}
\alias{retractcheck_docx}
\title{Check docx file for retractions}
\usage{
retractcheck_docx(path)
}
\arguments{
\item{path}{Path to DOCX file to check}
}
\value{
\code{\link{retractcheck}} dataframe without filenames
}
\description{
Check a DOCX file for retractions.
}
\examples{
\dontrun{
  retractcheck_docx('manuscript.docx')
}
}
