% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractcheck.R
\name{retractcheck_rtf}
\alias{retractcheck_rtf}
\title{Check rtf file for retractions}
\usage{
retractcheck_rtf(path)
}
\arguments{
\item{path}{Path to rtf file to check}
}
\value{
\code{\link{retractcheck}} dataframe
}
\description{
Check a rtf file for retractions.
}
\examples{
\dontrun{
  retractcheck_rtf('manuscript.rtf')
}
}
